backend/requirements.txt
# Discord
discord.py>=2.3.2
python-dotenv>=1.0.0
selenium>=4.15.2  # For LinkedIn scraping

# API
fastapi>=0.104.1
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9

# Utilities
requests>=2.31.0
aiohttp>=3.8.0
pydantic-settings>=2.0.0

# LinkedIn
linkedin-api>=0.1.0

# Cryptography
cryptography>=41.0.0 